Add tests for when the precision value registered for a type is of the wrong type.
